Fixed multi-threading
Pickling the method has been changed, now no methods or variables are getting pickled - engine had to be made a global object, and the UI is only running on the main thread
There is now an activate function, which we may no longr need, as the engine is a global object.
Next steps: Interrupt the system when detects speech, whether its ASR or a hardcoded value of our own. Also user and gpt speaking and lkistening at correct times - currenlty only does the first sentence